+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 更改日志 | Change Log
 
+## v0.3.14 / 2024-11-20
+
+- [v0.3.13 ... v0.3.14](https://github.com/siyuan-community/siyuan-sdk/compare/v0.3.13...v0.3.14)
 - 修复类型定义文件入口路径错误问题 | Fix the issue of incorrect entry path of type definition file
 - 调整 API `/api/system/getConf` | Adjust API `/api/system/getConf`
 - 调整 API `/api/repo/openRepoSnapshotDoc` | Adjust API `/api/repo/openRepoSnapshotDoc`